Technical audit fixes + BcInheritFrom positional path
Audit fixes:
- Fix integer division semantics: quot→div, rem→mod (Nix uses floored
  division, not truncated toward zero). Affects builtinDiv, builtinMod,
  and evalDiv.
- Fix nn_attrset.c: document partial-realloc-failure behavior, remove
  unused nn_attrset_intersect and nn_attrset_values_ptr.
- Fix nn_bytecode.c: ensure_op_space/ensure_data_space return error
  codes instead of failing silently; add overflow guard on capacity
  doubling.
- Fix nn_lambda.c: add NN_ASSERT bounds checks on entry accessors to
  prevent null deref when formal_count==0.
- Add nn_assert.h: debug-mode bounds checking macro (compiles to
  nothing under NDEBUG).
- Remove dead code: cattrsetIntersect FFI wrapper + C functions.

BcInheritFrom fix:
- allBcPositional, bcBindingSlotCount, buildBcSlotThunks, and
  buildBcAttrMapFromSlots now handle BcInheritFrom (inherit-from in
  positional let/rec blocks). The resolution pass treated inherit-from
  as positional but the evaluator did not, causing env slot count
  mismatch and NN_ASSERT idx-out-of-bounds on nixpkgs eval.

Stress test: 6.25 MB residency, 56.3% GC productivity (was 17.7 MB /
~40% after M5).

+++ src/Nix/Builtins.hs
@@ -15,6 +15,7 @@ module Nix.Builtins
   )
 where
 
+import Data.Int (Int64)
 import qualified Data.Map.Strict as Map
 import Data.Text (Text)
 import qualified Data.Text as T
@@ -34,7 +35,7 @@ import Nix.Store.Path (platformStoreDirText)
 --
 -- @searchPaths@ populates @builtins.nixPath@.  Parsed from @NIX_PATH@
 -- by 'parseNixPath'.  In tests, pass @[]@.
-builtinEnv :: Integer -> [Thunk] -> Env
+builtinEnv :: Int64 -> [Thunk] -> Env
 builtinEnv timestamp searchPaths =
   let scope =
         attrSetFromMap $
@@ -77,21 +78,21 @@ topLevelBuiltin name = (name, evaluated (VBuiltin name []))
 
 -- | Like 'builtinEnv' but with additional scope bindings overlaid on
 -- the top-level environment.  Used by @scopedImport@.
-builtinEnvWithScope :: Integer -> [Thunk] -> [(Text, Thunk)] -> Env
+builtinEnvWithScope :: Int64 -> [Thunk] -> [(Text, Thunk)] -> Env
 builtinEnvWithScope timestamp searchPaths scope =
   let base = builtinEnv timestamp searchPaths
       scopeMap = Map.fromList scope
    in newCEnv nullPtr 0 (Just (attrSetFromMap scopeMap)) (Just base) nullPtr 0
 
 -- | The @builtins@ attribute set, derived from the central registry.
-builtinsAttrSet :: Integer -> [Thunk] -> NixValue
+builtinsAttrSet :: Int64 -> [Thunk] -> NixValue
 builtinsAttrSet timestamp searchPaths =
   VAttrs $ attrSetFromMap $ Map.union builtinEntries (standardEntries timestamp searchPaths)
   where
     builtinEntries =
       Map.fromList [(name, evaluated (VBuiltin name [])) | name <- builtinNames]
 
-standardEntries :: Integer -> [Thunk] -> Map.Map Text Thunk
+standardEntries :: Int64 -> [Thunk] -> Map.Map Text Thunk
 standardEntries timestamp searchPaths =
   Map.fromList
     [ ("true", evaluated (VBool True)),
